Added email column to CommitData Entity, and gathered it during REST API call
